requirements.txt
Django==3.2.6
djangorestframework==3.12.4
mysqlclient==1.4.6
django-mysql==3.8.1

drf-yasg2==1.19.4

PyJWT==1.7.1
django-cors-headers


